+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/bean/JpaAuditingConfiguration.java
@@ -1,0 +1,9 @@
+package sellyourunhappiness.api.config.bean;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaAuditingConfiguration {
+}
