+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/bean/ThreadPoolTaskExecutorConfig.java
@@ -1,0 +1,18 @@
+package sellyourunhappiness.api.config.bean;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.task.TaskExecutor;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+@Configuration
+public class ThreadPoolTaskExecutorConfig {
+	@Bean
+	public TaskExecutor threadPoolTaskExecutor() {
+		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+		executor.setCorePoolSize(5);
+		executor.setMaxPoolSize(5);
+		executor.initialize();
+		return executor;
+	}
+}
