+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/security/handler/CustomOAuth2LoginFailureHandler.java
@@ -1,0 +1,19 @@
+package sellyourunhappiness.api.config.security.handler;
+
+import java.io.IOException;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Component
+public class CustomOAuth2LoginFailureHandler implements AuthenticationFailureHandler {
+	@Override
+	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
+		response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+		response.getWriter().write("소셜 로그인 실패, 서버 로그를 확인해주세요.");
+	}
+}
