+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/security/oauth2/social/GoogleOAuth2UserInfo.java
@@ -1,0 +1,26 @@
+package sellyourunhappiness.api.config.security.oauth2.social;
+
+import java.util.Map;
+
+import sellyourunhappiness.api.config.security.oauth2.social.abstarcts.OAuth2UserInfo;
+
+public class GoogleOAuth2UserInfo extends OAuth2UserInfo {
+	public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
+		super(attributes);
+	}
+
+	@Override
+	public String getEmail() {
+		return (String)attributes.get("email");
+	}
+
+	@Override
+	public String getNickname() {
+		return (String)attributes.get("name");
+	}
+
+	@Override
+	public String getImageUrl() {
+		return (String)attributes.get("picture");
+	}
+}
