+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/security/oauth2/social/abstarcts/OAuth2UserInfo.java
@@ -1,0 +1,19 @@
+package sellyourunhappiness.api.config.security.oauth2.social.abstarcts;
+
+import java.util.Map;
+
+
+public abstract class OAuth2UserInfo {
+
+	protected Map<String, Object> attributes;
+
+	public OAuth2UserInfo(Map<String, Object> attributes) {
+		this.attributes = attributes;
+	}
+
+	public abstract String getEmail();
+
+	public abstract String getNickname();
+
+	public abstract String getImageUrl();
+}
