+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/slack/annotation/SlackNotification.java
@@ -1,0 +1,11 @@
+package sellyourunhappiness.api.config.slack.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SlackNotification {
+}
