+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/config/slack/config/SlackConfig.java
@@ -1,0 +1,18 @@
+package sellyourunhappiness.api.config.slack.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import net.gpedro.integrations.slack.SlackApi;
+
+@Configuration
+public class SlackConfig {
+	@Value("${sellyourunhappiness.slack.token}")
+	private String token;
+
+	@Bean
+	public SlackApi slackApi() {
+		return new SlackApi("https://hooks.slack.com/services/" + token);
+	}
+}
