+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/user/application/UserBroker.java
@@ -5,35 +5,35 @@ import java.util.Map;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import sellyourunhappiness.core.user.application.JwtService;
-import sellyourunhappiness.global.config.security.oauth2.OAuthAttributes;
+import lombok.extern.slf4j.Slf4j;
+import sellyourunhappiness.api.config.jwt.application.JwtService;
+import sellyourunhappiness.api.user.dto.UserResponse;
 import sellyourunhappiness.core.user.application.UserService;
 import sellyourunhappiness.core.user.domain.User;
 import sellyourunhappiness.core.user.domain.enums.SocialType;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class UserBroker {
 	private final UserService userService;
 	private final JwtService jwtService;
 
-	public User getUserByEmail(String email) {
-		return userService.findByEmail(email)
-			.orElseThrow(() -> new IllegalArgumentException("유저를 찾을 수 없습니다.: " + email));
+	public UserResponse getUserByEmail(String email) {
+		User user = userService.findByEmail(email);
+		return new UserResponse(user.getName(), user.getEmail());
 	}
 
 	public Map<String, String> refreshAccessToken(String refreshToken) {
 		return jwtService.refreshAccessToken(refreshToken);
 	}
 
-	public User getUser(OAuthAttributes attributes, SocialType socialType) {
-		return userService.findBySocialTypeAndEmail(socialType, attributes.getOauth2UserInfo().getEmail())
-			.orElseGet(() -> {
-				String name = attributes.getOauth2UserInfo().getNickname();
-				String email = attributes.getOauth2UserInfo().getEmail();
-				String profileUrl = attributes.getOauth2UserInfo().getImageUrl();
-
-				return userService.saveUser(name, email, profileUrl, socialType);
-			});
+	public User getUserByOauth(String name, String email, String profileUrl, SocialType socialType) {
+		try {
+			return userService.findBySocialTypeAndEmail(socialType, email);
+		} catch (IllegalArgumentException e) {
+			log.debug(e.getMessage());
+		}
+		return userService.save(name, email, profileUrl, socialType);
 	}
 }
