+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/api/user/dto/UserResponse.java
@@ -1,0 +1,14 @@
+package sellyourunhappiness.api.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UserResponse {
+	private String username;
+	private String email;
+
+	public UserResponse(String username, String email) {
+		this.username = username;
+		this.email = email;
+	}
+}
