--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/JpaAuditingConfiguration.java
@@ -1,9 +1,0 @@
-package sellyourunhappiness.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaAuditingConfiguration {
-}
