--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/ThreadPoolTaskExecutorConfig.java
@@ -1,16 +1,0 @@
-package sellyourunhappiness.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.core.task.TaskExecutor;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-public class ThreadPoolTaskExecutorConfig {
-	@Bean
-	public TaskExecutor threadPoolTaskExecutor() {
-		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-		executor.setCorePoolSize(5);
-		executor.setMaxPoolSize(5);
-		executor.initialize();
-		return executor;
-	}
-}
