--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/security/handler/CustomOAuth2LoginFailureHandler.java
@@ -1,24 +1,0 @@
-package sellyourunhappiness.global.config.security.handler;
-
-import java.io.IOException;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Component
-public class CustomOAuth2LoginFailureHandler implements AuthenticationFailureHandler {
-	@Override
-	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
-		AuthenticationException exception) throws
-		ServletException, IOException {
-		response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-		response.getWriter().write("소셜 로그인 실패, 서버 로그를 확인해주세요.");
-	}
-}
