--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/security/oauth2/abstracts/OAuth2UserInfo.java
@@ -1,21 +1,0 @@
-package sellyourunhappiness.global.config.security.oauth2.abstracts;
-
-import java.util.Map;
-
-
-public abstract class OAuth2UserInfo {
-
-	protected Map<String, Object> attributes;
-
-	public OAuth2UserInfo(Map<String, Object> attributes) {
-		this.attributes = attributes;
-	}
-
-	public abstract String getId();
-
-	public abstract String getEmail();
-
-	public abstract String getNickname();
-
-	public abstract String getImageUrl();
-}
