--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/mixin/BaseTimeMixin.java
@@ -1,23 +1,0 @@
-package sellyourunhappiness.global.mixin;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseTimeMixin {
-	@CreatedDate
-	private LocalDateTime createdDate;
-
-	@LastModifiedDate
-	private LocalDateTime modifiedDate;
-
-}
