--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/response/ErrorResponse.java
@@ -1,14 +1,0 @@
-package sellyourunhappiness.global.response;
-
-import org.springframework.http.HttpStatus;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class ErrorResponse {
-
-	private final HttpStatus status;
-	private final String message;
-}
