--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/slack/annotation/SlackNotification.java
@@ -1,11 +1,0 @@
-package sellyourunhappiness.global.slack.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SlackNotification {
-}
