--- sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/slack/config/SlackConfig.java
@@ -1,19 +1,0 @@
-package sellyourunhappiness.global.slack.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import net.gpedro.integrations.slack.SlackApi;
-
-@Configuration
-public class SlackConfig {
-
-	@Value("${sellyourunhappiness.slack.token}")
-	private String token;
-
-	@Bean
-	public SlackApi slackApi() {
-		return new SlackApi("https://hooks.slack.com/services/" + token);
-	}
-}
