+++ sellyourunhappiness-core/src/main/java/sellyourunhappiness/core/user/application/UserService.java
@@ -1,40 +1,31 @@
 package sellyourunhappiness.core.user.application;
 
-import java.util.Optional;
+import static sellyourunhappiness.core.user.domain.User.*;
 
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import sellyourunhappiness.core.user.domain.User;
-import sellyourunhappiness.core.user.domain.enums.Role;
 import sellyourunhappiness.core.user.domain.enums.SocialType;
-import sellyourunhappiness.core.user.domain.enums.UserStatus;
 import sellyourunhappiness.core.user.infrastructure.UserRepository;
-
-
 @RequiredArgsConstructor
 @Service
 public class UserService {
 	private final UserRepository userRepository;
 
-	public Optional<User> findByEmail(String email) {
-		return userRepository.findByEmail(email);
+	public User findByEmail(String email) {
+		return userRepository.findByEmail(email)
+			.orElseThrow(() -> new IllegalArgumentException("유저를 찾을 수 없습니다.: " + email));
 	}
 
-	public User saveUser(String name, String email, String profileUrl, SocialType socialType) {
-		User createdUser = User.builder()
-			.name(name)
-			.email(email)
-			.profileURL(profileUrl)
-			.role(Role.USER)
-			.status(UserStatus.ACTIVE)
-			.socialType(socialType)
-			.build();
+	public User save(String name, String email, String profileUrl, SocialType socialType) {
+		User createdUser = create(name, email, profileUrl, socialType);
 		return userRepository.save(createdUser);
 	}
 
-	public Optional<User> findBySocialTypeAndEmail(SocialType socialType, String email) {
-		return userRepository.findBySocialTypeAndEmail(socialType, email);
+	public User findBySocialTypeAndEmail(SocialType socialType, String email) {
+		return userRepository.findBySocialTypeAndEmail(socialType, email)
+			.orElseThrow(() -> new IllegalArgumentException("소셜 타입, 이메일과 일치하는 유저가 없습니다. SocialType : " + socialType + " email : " + email));
 	}
 
 }
